djangoapi/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.2.1
djangorestframework==3.14.0
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==18.7.0
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
Pygments==2.15.1
pytest==7.3.1
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
